Update
The layouts of the main screen were designed. Edit Text was added, which allows the user to add their name, and a plus button was added to save the user's name in the SQL database. By using the recycler view, the names of the users have been adjusted to appear in the middle of the main screen, and a minus button has been added if the user wishes to remove any user. This minus button serves to delete the name of the user who wants to be deleted from the SQL database.