About the return button
First, I added the selectchoice activity. I opened this activity so that the user can choose the desired game. I also wrote the code required to return to the previous screen when the user presses the back button, and I put a return button in the upper left corner in case the user wants to return to the user adding menu from the screen where he can select the games.